async/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.8
more-itertools==10.4.0
multidict==6.0.5
requests==2.32.3
SQLAlchemy==2.0.34
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.8
python-dotenv
